Display Image in recyclerview

+++ app/src/main/java/com/example/camapp/ImageModel.kt
@@ -2,4 +2,4 @@ package com.example.camapp
 
 import android.net.Uri
 
-data class ImageModel(val uri: Uri,val imgLat: Long, val imgLong: Long )
+data class ImageModel(val uri: Uri, val imgLat: Double, val imgLong: Double)
